only store error codes for single change proto
