+++ modules/core/src/main/scala/vinyldns/core/domain/batch/SingleChange.scala
@@ -18,7 +18,7 @@ package vinyldns.core.domain.batch
 
 import java.util.UUID
 
-import vinyldns.core.domain.DomainValidationError
+import vinyldns.core.domain.DomainValidationErrorType.DomainValidationErrorType
 import vinyldns.core.domain.batch.SingleChangeStatus.SingleChangeStatus
 import vinyldns.core.domain.record.RecordData
 import vinyldns.core.domain.record.RecordType.RecordType
@@ -34,7 +34,7 @@ sealed trait SingleChange {
   val typ: RecordType
   val inputName: String
   val zoneName: Option[String]
-  val validationErrors: List[DomainValidationError]
+  val validationErrors: List[DomainValidationErrorType]
   val recordKey: Option[RecordKey] = (zoneId, recordName, typ) match {
     case (Some(zid), Some(rname), t) => Some(RecordKey(zid, rname, t))
     case _ => None
@@ -92,7 +92,7 @@ final case class SingleAddChange(
     systemMessage: Option[String],
     recordChangeId: Option[String],
     recordSetId: Option[String],
-    validationErrors: List[DomainValidationError] = List.empty,
+    validationErrors: List[DomainValidationErrorType] = List.empty,
     id: String = UUID.randomUUID().toString)
     extends SingleChange
 
@@ -106,7 +106,7 @@ final case class SingleDeleteChange(
     systemMessage: Option[String],
     recordChangeId: Option[String],
     recordSetId: Option[String],
-    validationErrors: List[DomainValidationError] = List.empty,
+    validationErrors: List[DomainValidationErrorType] = List.empty,
     id: String = UUID.randomUUID().toString)
     extends SingleChange
